Keyboard Selection Implementation
The keys 'a' and 'd' now change what related video is selected as
denoted by a white border.  The border settings still need to be changed
to make it look better.

I made a slight change to the json file so that the order that the
related videos would be displayed would match the order of the videos in
the related array.

I attempted to implement a 'w' command to play the video selected but I
couldn't figure out how to do the loadVideo action from a function.